requirements.txt
# Core Framework
flask==3.0.0
gunicorn==21.2.0

# AI & API
anthropic>=0.40.0

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
trafilatura==1.8.0

# Data Processing
pandas>=2.2.2
openpyxl==3.1.2

# Database (Production)
psycopg2-binary==2.9.9

# Configuration
python-dotenv>=1.0.1

# Optional: Authentication
# clerk-backend-api==0.1.0

# Optional: Email
# resend==0.7.0

# Security
flask-cors==4.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
